2023/02/04-0839

+++ StartCheck.py
@@ -22,6 +22,7 @@ import json
 import glob
 import shutil
 from tkinter import filedialog
+from tkinter import messagebox
 from CheckTool import CheckTool
 import logging
 import threading
@@ -30,13 +31,18 @@ from datetime import datetime
 # グリーバル変数の定義
 time_sta =  0       # 経過時間を表示するための開始時刻
 flag1 = False       # 表示ウインドウを制御するための変数（Trueの場合は表示、Falseは表示終了）
+ErrorFlag = False   # エラーの有無
+ErrorMessage = ""   # エラーメッセージ
 fname = ""          # 現在処理中のファイル名
 folderName = ""     # 現在処理中のファイルのあるフォルダー名
 dir1 = ""           # 処理前フォルダー
 dir2 = ""           # 処理後フォルダー
 dir3 = ""           # ログのフォルダー
 dir4 = ""           # パラメータファイルのテンプレートのフォルダー
+dir5 = ""           # エラーデータのフォルダー
 paraFileName = "para.json" # パラメータファイルの名称
+runLogFile = "処理結果ログ.txt"
+systemLogFile = "system.log"
 
 #============================================================================
 #  作業フォルダーの設定データ（init.json）読込
@@ -47,44 +53,66 @@ paraFileName = "para.json" # パラメータファイルの名称
 #============================================================================
 
 def CreateFolfer():
-    global dir1, dir2, dir3, dir4, folderName, paraFileName
+    global flag1, fname, dir1, dir2, dir3, dir4, dir5, folderName, paraFileName
+    global ErrorFlag, ErrorMessage, runLogFile, systemLogFile
     try:
         # CalcNames = [["SS7", "CheckTool"], ["その他", "CheckTool"]]
         initFile = "init.json"
         if not os.path.isfile(initFile):
-            dir = os.getcwd()
-            fld = filedialog.askdirectory(initialdir=dir)
-            dir1 = fld + "/処理前フォルダ"
-            dir2 = fld + "/処理後フォルダ"
-            dir3 = fld + "/ログ"
-            dir4 = fld + "/パラメータファイルのテンプレート"
-            if not os.path.isdir(dir1):
-                os.mkdir(dir1)
-            if not os.path.isdir(dir2):
-                os.mkdir(dir2)
-            if not os.path.isdir(dir3):
-                os.mkdir(dir3)
-            if not os.path.isdir(dir4):
-                os.mkdir(dir4)
+            ret = messagebox.askyesno('確認', '作業フォルダの設定ファイルがありません。\n作業フォルダの設定を行いますか？')
+            if ret == True:
+                dir = os.getcwd()
+                fld = filedialog.askdirectory(initialdir=dir)
+                dir1 = fld + "/処理前フォルダ"
+                dir2 = fld + "/処理後フォルダ"
+                dir3 = fld + "/ログ"
+                dir4 = fld + "/パラメータファイルのテンプレート"
+                dir5 = fld + "/エラーフォルダ"
+                if not os.path.isdir(dir1):
+                    os.mkdir(dir1)
+                if not os.path.isdir(dir2):
+                    os.mkdir(dir2)
+                if not os.path.isdir(dir3):
+                    os.mkdir(dir3)
+                if not os.path.isdir(dir4):
+                    os.mkdir(dir4)
+                if not os.path.isdir(dir5):
+                    os.mkdir(dir5)
 
-            pageData = {"処理前フォルダ": dir1, "処理後フォルダ": dir2, "ログ": dir3, "パラメータファイルのテンプレート": dir4}
-            # data_json = json.dumps(pageData, indent=4, ensure_ascii=False)
-            with open('init.json', 'w', encoding="utf-8") as fp:
-                json.dump(pageData, fp, indent=4, ensure_ascii=False)
-                fp.close()
 
-            para = {"数値の閾値": 0.95, "開始ページ": 2, "終了ページ": 0}
-            with open(dir4+'/'+paraFileName, 'w', encoding="utf-8") as fp:
-                json.dump(para, fp, indent=4, ensure_ascii=False)
-                fp.close()
+                pageData = {"処理前フォルダ": dir1, "処理後フォルダ": dir2, "ログ": dir3,
+                                "パラメータファイルのテンプレート": dir4, "エラーフォルダ": dir5}
+                # data_json = json.dumps(pageData, indent=4, ensure_ascii=False)
+                with open('init.json', 'w', encoding="utf-8") as fp:
+                    json.dump(pageData, fp, indent=4, ensure_ascii=False)
+                    fp.close()
+
+                para = {"数値の閾値": 0.95, "開始ページ": 2, "終了ページ": 0}
+                with open(dir4+'/'+paraFileName, 'w', encoding="utf-8") as fp:
+                    json.dump(para, fp, indent=4, ensure_ascii=False)
+                    fp.close()
+
+                if not os.path.isfile(dir5+'/'+runLogFile):
+                    Message = "実行結果のログファイルの初期設定"
+                    AddLog(Message)
+
+                flag1 = False
+                return False
+                
+            else:
+                flag1 = False
+                return False
+
 
         else:
+            
             json_open = open(initFile, 'r', encoding="utf-8")
             json_load = json.load(json_open)
             dir1 = json_load['処理前フォルダ']
             dir2 = json_load['処理後フォルダ']
             dir3 = json_load['ログ']
             dir4 = json_load['パラメータファイルのテンプレート']
+            dir5 = json_load['エラーフォルダ']
             if not os.path.isdir(dir1):
                 os.mkdir(dir1)
             if not os.path.isdir(dir2):
@@ -92,24 +120,48 @@ def CreateFolfer():
             if not os.path.isdir(dir3):
                 os.mkdir(dir3)
             if not os.path.isdir(dir4):
-                os.mkdir(dir4)
-            
+                os.mkdir(dir4)          
+            if not os.path.isdir(dir5):
+                os.mkdir(dir5)          
             json_open.close()
-            
 
             if not os.path.isfile(dir4+'/'+paraFileName):
                 para = {"数値の閾値": 0.95, "開始ページ": 2, "終了ページ": 0}
-                with open(dir4+'/'+paraFileName, 'w') as fp:
-                    json.dump(para, fp, indent=4, ensure_ascii=False)
-                    fp.close()
+                try:
+                    with open(dir4+'/'+paraFileName, 'w') as fp:
+                        json.dump(para, fp, indent=4, ensure_ascii=False)
+                        fp.close()
+                except OSError as e:
+                    print(e)
+                    logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+                    ErrorMessage += "パラメータファイルの読込エラー\n"
+                    ErrorFlag = True
+                    flag1 = False
+                except:
+                    print("")
+                    logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+                    ErrorMessage += "パラメータファイルの読込エラー\n"
+                    ErrorFlag = True
+                    flag1 = False
+            
+            flag1 = True
+                
+            return True
 
     except OSError as e:
         print(e)
         logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
-        
+        ErrorMessage += "システムエラー\n"
+        ErrorFlag = True
+        flag1 = False
+        return False
     except:
         print("")
         logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+        ErrorMessage += "原因不明のエラー\n"
+        ErrorFlag = True
+        flag1 = False
+        return False
     #*********************************************************************************
 
 
@@ -120,7 +172,8 @@ def CreateFolfer():
 
 def RunCheck():
     global time_sta
-    global flag1, fname, dir1, dir2, dir3, dir4, folderName, paraFileName
+    global flag1, fname, dir1, dir2, dir3, dir4, dir5, folderName, paraFileName
+    global ErrorFlag, ErrorMessage, runLogFile, systemLogFile
     
     try:
 
@@ -137,6 +190,7 @@ def RunCheck():
         # stpage = 242
         # edpage = 250
         if len(folders) > 0:
+            AddLog("処理の開始")
             for folder in folders:      # フォルダー毎に処理を実行
                 if not "検出結果" in folder:  # フォルダー名に"検出結果"が含まれる場合は結果フォルダなので無視する。
                     folderName = folder     # 表示ウィンドウに表示させるフォルダー名
@@ -174,7 +228,8 @@ def RunCheck():
                                     # 検査がエラーなく終了した場合の処理
                                     # 処理後フォルダーに同じ名称のデータフォルダーがある場合は、上書きせずに、
                                     # データフォルダー名に'(n)'を追加して移動
-
+                                    message = folderName + "/" + fname + ":数値の検出処理OK"
+                                    AddLog(message)
                                     # フォルダー名の最後の3文字が (n) の場合は何番目であるか
                                     t1 = outfolder[len(outfolder)-3:]
                                     if t1[0] == "(" and t1[len(t1)-1] == ")" :
@@ -197,8 +252,16 @@ def RunCheck():
                                             if not os.path.isdir(newFolder):
                                                 new_path = shutil.move(path1, newFolder)
                                                 break
+
+                                    message = folderName + "/" + fname + ":フォルダの移動処理OK"
+                                    AddLog(message)
                                     
-                                        
+                    else:
+                        ErrorMessage += folderName + "にPDFファイルがありません\n"
+                        ErrorFlag = True
+                        flag1 = False
+
+            AddLog("処理の終了")    
     
         t1 = time.time() - time_sta
         print("time = {} sec".format(t1))
@@ -207,11 +270,15 @@ def RunCheck():
     except OSError as e:
         print(e)
         logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+        ErrorMessage += "システムエラー\n"
+        ErrorFlag = True
         flag1 = False
         
     except:
         print("")
         logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+        ErrorMessage += "原因不明のエラー\n"
+        ErrorFlag = True
         flag1 = False
     #*********************************************************************************
 
@@ -223,73 +290,113 @@ def RunCheck():
 
 def main():
     global time_sta
-    global flag1, fname, dir1, dir2, dir3, dir4, folderName
+    global flag1, fname, dir1, dir2, dir3, dir4, dir5, folderName, paraFileName
+    global ErrorFlag, ErrorMessage, runLogFile, systemLogFile
+    
+    if CreateFolfer():
+        los_file = dir3 + "/" + systemLogFile
+        # log 出力レベルの設定
+        logging.basicConfig(filename=los_file,level=logging.WARNING,
+                    format="%(asctime)s %(levelname)s %(message)s")
+        logging.debug('debug')
+        logging.info('info')
+        logging.warning('warnig')
+        logging.error('error')
+        logging.critical('critical')
 
-    CreateFolfer()
+        time_sta = time.time()  # 開始時刻の記録
 
-    los_file = dir3 + "/" + "CheckTool.log"
-    # log 出力レベルの設定
-    logging.basicConfig(filename=los_file,level=logging.WARNING,
-                format="%(asctime)s %(levelname)s %(message)s")
-    logging.debug('debug')
-    logging.info('info')
-    logging.warning('warnig')
-    logging.error('error')
-    logging.critical('critical')
+        root = tk.Tk()
+        # root = tk.Toplevel()
+    
+        Width = 800
+        Height = 600
+        root.title(u"Calculation Cheet Check")
+        root.geometry("800x600")
+        root.geometry("{}x{}".format(Width,Height))
 
-    time_sta = time.time()  # 開始時刻の記録
+        Static1 = tk.Label(text=u'\n構造計算書の数値検索プログラム', font=("MSゴシック", "28", "bold"))
+        Static1.pack()
 
-    root = tk.Tk()
-    Width = 800
-    Height = 600
-    root.title(u"Calculation Cheet Check")
-    root.geometry("800x600")
-    root.geometry("{}x{}".format(Width,Height))
+        Static2 = tk.Label(text=u'\n一般財団法人日本建築総合試験所', font=("MSゴシック", "28", "bold"))
+        Static2.pack()
 
-    Static1 = tk.Label(text=u'\n構造計算書の数値検索プログラム', font=("MSゴシック", "28", "bold"))
-    Static1.pack()
+        Static3 = tk.Label(text=u'\n\nファイル名：\n\nファイル名：', font=("MSゴシック", "28", "bold"))
+        Static3.pack()
 
-    Static2 = tk.Label(text=u'\n一般財団法人日本建築総合試験所', font=("MSゴシック", "28", "bold"))
-    Static2.pack()
+        Static4 = tk.Label(text=u'\n経過時間：', font=("ヒラギノ角ゴシック", "28", "bold"))
+        Static4.pack()
 
-    Static3 = tk.Label(text=u'\n\nファイル名：\n\nファイル名：', font=("MSゴシック", "28", "bold"))
-    Static3.pack()
+        root.update_idletasks()
+        ww=root.winfo_screenwidth()
+        lw=root.winfo_width()
+        wh=root.winfo_screenheight()
+        lh=root.winfo_height()
+        # canvas=tk.Canvas(root,width=lw,heigh=lh)
+        # canvas.pack()#ここを書かないとcanvasがうまく入らない．
 
-    Static4 = tk.Label(text=u'\n経過時間：', font=("ヒラギノ角ゴシック", "28", "bold"))
-    Static4.pack()
+        root.geometry(str(lw)+"x"+str(lh)+"+"+str(int(ww/2-lw/2))+"+"+str(int(wh/2-lh/2)) )
 
-    root.update_idletasks()
-    ww=root.winfo_screenwidth()
-    lw=root.winfo_width()
-    wh=root.winfo_screenheight()
-    lh=root.winfo_height()
-    # canvas=tk.Canvas(root,width=lw,heigh=lh)
-    # canvas.pack()#ここを書かないとcanvasがうまく入らない．
+        thread1 = threading.Thread(target=RunCheck)
+        thread1.start()
+        flag1 = True
+        ErrorFlag = False
+        ErrorMessage = ""
+        # i = 0
+        while flag1:
+            root.update()
+            # now_h=datetime.now().hour
+            # now_s=datetime.now().second
+            # now_m=datetime.now().minute
+            # now_time=str(now_h)+":"+str(now_m)+":"+str(now_s)
+            Static3["text"] = '\n\nフォルダー名：' + folderName + '\n\nファイル名：' + fname
+            Static4["text"] = "\n経過時間：{:7.0f}秒".format(time.time() - time_sta)
+            # canvas.create_text(lw/2,200,text=now_time,font=("",25,""),tag='Y') #タグを入れることで更新できるようにする．
+            # canvas.update()
+            # canvas.delete('Y')
+            # i += 1
+            # Static2["text"] = u'\n一般財団法人日本建築総合試験所{}'.format(i)
 
-    root.geometry(str(lw)+"x"+str(lh)+"+"+str(int(ww/2-lw/2))+"+"+str(int(wh/2-lh/2)) )
-
-    thread1 = threading.Thread(target=RunCheck)
-    thread1.start()
-    flag1 = True
-    # i = 0
-    while flag1:
-        root.update()
-        # now_h=datetime.now().hour
-        # now_s=datetime.now().second
-        # now_m=datetime.now().minute
-        # now_time=str(now_h)+":"+str(now_m)+":"+str(now_s)
-        Static3["text"] = '\n\nフォルダー名：' + folderName + '\n\nファイル名：' + fname
-        Static4["text"] = "\n経過時間：{:7.0f}秒".format(time.time() - time_sta)
-        # canvas.create_text(lw/2,200,text=now_time,font=("",25,""),tag='Y') #タグを入れることで更新できるようにする．
-        # canvas.update()
-        # canvas.delete('Y')
-        # i += 1
-        # Static2["text"] = u'\n一般財団法人日本建築総合試験所{}'.format(i)
-
-        time.sleep(1.0)
+            time.sleep(1.0)
+        
+        if ErrorFlag:
+            AddLog(ErrorMessage)
+            messagebox.showerror('エラー', ErrorMessage)
+    else:
+        return
 
     #*********************************************************************************
 
+def AddLog(Message1):
+    global flag1, fname, dir1, dir2, dir3, dir4, dir5, folderName, paraFileName
+    global ErrorFlag, ErrorMessage, runLogFile, systemLogFile
+    
+    if Message1 != "":
+        now = datetime.now()
+        Message = now.strftime('%Y/%m/%d %H:%M:%S')+ ":" +Message1
+        try:
+            if os.path.isfile(dir3+'/'+runLogFile):
+                with open(dir3+'/'+runLogFile, 'a', encoding="utf-8") as fp:
+                    print(Message, file=fp)
+                    fp.close()
+            else:
+                with open(dir3+'/'+runLogFile, 'w', encoding="utf-8") as fp:
+                    print(Message, file=fp)
+                    fp.close()
+
+        except OSError as e:
+            print(e)
+            logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+            ErrorMessage += "システムエラー\n"
+            ErrorFlag = True
+            flag1 = False
+            
+        except:
+            print("")
+            logging.exception(sys.exc_info())#エラーをlog.txtに書き込む
+            ErrorMessage += "原因不明のエラー\n"
+            ErrorFlag = True
+            flag1 = False
 
 
     #============================================================================
